### 2022-01-26 b5ced20
Adding support for periodic BCs
