+++ src/Geometry.jl
@@ -99,23 +99,60 @@ end
 # CartesianDiscreteModel
 
 function Geometry.CartesianDiscreteModel(
-  parts::AbstractPData{<:Integer},args...;kwargs...)
+  parts::AbstractPData{<:Integer},args...;isperiodic=map(i->false,size(parts)),kwargs...)
 
-  desc = CartesianDescriptor(args...; kwargs...)
+  desc = CartesianDescriptor(args...;isperiodic=isperiodic,kwargs...)
   nc = desc.partition
   msg = """
   A CartesianDiscreteModel needs a Cartesian subdomain partition
   of the right dimensions.
   """
   @assert length(size(parts)) == length(nc) msg
-  gcids = PCartesianIndices(parts,nc,PArrays.with_ghost)
-  gids = PRange(parts,nc,PArrays.with_ghost)
-  models = map_parts(parts,gcids) do part, gcids
-    cmin = first(gcids)
-    cmax = last(gcids)
-    CartesianDiscreteModel(desc,cmin,cmax)
+
+  if any(isperiodic)
+    model = _cartesian_model_with_periodic_bcs(parts,desc)
+  else
+    gcids = PCartesianIndices(parts,nc,PArrays.with_ghost)
+    models = map_parts(parts,gcids) do part, gcids
+      cmin = first(gcids)
+      cmax = last(gcids)
+      CartesianDiscreteModel(desc,cmin,cmax)
+    end
+    gids = PRange(parts,nc,PArrays.with_ghost)
+    model = DistributedDiscreteModel(models,gids)
   end
-  DistributedDiscreteModel(models,gids)
+  model
+end
+
+function _cartesian_model_with_periodic_bcs(parts,desc)
+  h = desc.sizes
+  _origin = map(desc.isperiodic,Tuple(desc.origin),h,size(parts)) do isp,o,h,np
+    isp&&np!=1 ? o-h : o
+  end
+  _sizes = h
+  _partition = map(desc.isperiodic,desc.partition,size(parts)) do isp,o,np
+    isp&&np!=1 ? o+2 : o
+  end
+   # Important, the map should be periodic if you want to integrate
+   # functions in the ghost cells.
+  _map = desc.map
+  _isperiodic = map(desc.isperiodic,size(parts)) do isp,np
+    np==1 ? isp : false
+  end
+  _desc = CartesianDescriptor(Point(_origin),_sizes,_partition;map=_map,isperiodic=_isperiodic)
+  isperiodic_global = map(desc.isperiodic,size(parts)) do isp,np
+    np==1 ? false : isp
+  end
+  in_bounds = Val(false)
+  gcids = PCartesianIndices(parts,desc.partition,PArrays.with_ghost,isperiodic_global,in_bounds)
+  nparts = size(parts)
+  models = map_parts(parts,gcids) do part, gcids
+    cmin = CartesianIndex(map((p,i,n)->( p&&n!=1 ? i+1 : i),desc.isperiodic,Tuple(first(gcids)),nparts))
+    cmax = CartesianIndex(map((p,i,n)->( p&&n!=1 ? i+1 : i),desc.isperiodic,Tuple(last(gcids)),nparts))
+    CartesianDiscreteModel(_desc,cmin,cmax)
+  end
+  gids = PRange(parts,desc.partition,PArrays.with_ghost,isperiodic_global)
+  model = DistributedDiscreteModel(models,gids)
 end
 
 ## Helpers to partition a serial model
