### 2022-01-27 4a1216d
More work on periodic BCs (test are still pending)
