### 2022-01-31 #4085 @Kevin-Mattheus-Moerman
@jedbrown @PetrKryslUCSD,  @Leticia-maria, @TeroFrondelius, @KristofferC, @rouson, @tjhei, @chrisrichardson, @Kevin-Mattheus-Moerman, @vijaysm, @likask would you be interesting in reviewing this software: "GridapDistributed: a massively parallel finite element toolbox in Julia" for [JOSS](https://joss.theoj.org)?

Short paper: https://github.com/openjournals/joss-papers/blob/joss.04085/joss.04085/10.21105.joss.04085.pdf
Software: https://github.com/gridap/GridapDistributed.jl/
