### 2022-02-03 #4085 @Kevin-Mattheus-Moerman
@Leticia-maria we used to aim for about 2-3 weeks but we can be (and have to be) quite flexible as we've seen (and experienced personally) that most people are struggling to keep up with things during the pandemic. So if you could aim for ~3 weeks great, but if you need more time just let me know. Note we have not formally started the review yet since this is the pre-review issue. You may start already if you wish but you can also wait until we've opened the review issue.
