+++ README.md
@@ -9,13 +9,19 @@ Parallel distributed-memory version of `Gridap.jl`.
 
 ## Purpose
 
-`GridapDistributed.jl` provides fully-parallel distributed memory data structures and associated methods for the Finite Element (FE) numerical solution of Partial Differential Equations (PDEs) on parallel computers, from multi-core CPU desktop computers, to HPC clusters and supercomputers. These distributed data structures are designed to mirror as far as possible their counterparts in the [`Gridap.jl`](https://github.com/gridap/Gridap.jl) software package, while implementing/leveraging most of their abstract interfaces. As a result, sequential Julia scripts written in the high level API of `Gridap.jl` can be used almost verbatim up to minor adjustments in a parallel context using `GridapDistributed.jl`. This is indeed one of the main advantages of `GridapDistributed.jl` and a major design goal that we pursue. 
+`GridapDistributed.jl` provides a fully-parallel distributed memory extension of the [`Gridap.jl`](https://github.com/gridap/Gridap.jl) library. It allows users to approximate PDEs on parallel computers, from multi-core CPU desktop computers to HPC clusters and supercomputers. The sub-package is designed to be as non-intrusive as possible. As a result, sequential Julia scripts written in the high level API of `Gridap.jl` can be used almost verbatim up to minor adjustments in a parallel context using `GridapDistributed.jl`. 
 
-At present, `GridapDistributed.jl` provides scalable parallel data structures for grid handling,  finite element spaces setup, and distributed linear system assembly. For the latter part, i.e., global distributed sparse matrices and vectors, `GridapDistributed.jl` relies on [`PartitionedArrays.jl`](https://github.com/fverdugo/PartitionedArrays.jl) as distributed linear algebra backend. This implies, among others, that all `GridapDistributed.jl` driver programs can be either run in sequential execution mode--very useful for developing/debugging parallel programs--, see `test/sequential/` folder for examples, or in message-passing (MPI) execution mode--when you want to deploy the code in the actual parallel computer and perform a fast simulation., see `test/mpi/` folder for examples.
+At present, `GridapDistributed.jl` provides scalable parallel data structures for grid handling,  finite element spaces setup, and distributed linear system assembly. For the latter part, i.e., global distributed sparse matrices and vectors, `GridapDistributed.jl` relies on [`PartitionedArrays.jl`](https://github.com/fverdugo/PartitionedArrays.jl) as distributed linear algebra backend. 
+
+## Documentation
+
+Given that GridapDistributed.jl and Gridap.jl share almost the same high-level API. We refer to the documentation of `Gridap.jl` for more details about the API. In the example below, we show the minor differences among the APIs of `Gridap.jl` and `GridapDistributed.jl`. We refer to the `test/` folder for additional `GridapDistributed.jl` examples.
+
+`GridapDistributed.jl` driver programs can be either run in sequential execution mode--very useful for developing/debugging parallel programs--, see `test/sequential/` folder for examples, or in message-passing (MPI) execution mode--when you want to deploy the code in the actual parallel computer and perform a fast simulation., see `test/mpi/` folder for examples.
 
 ## Example
 
-The abovementioned design goal in action can be observed in the following example Julia code snippet 
+The following Julia code snippet solves a 2D Poisson problem in parallel on the unit square
 
 ```julia
 using Gridap
@@ -41,17 +47,12 @@ prun(mpi,partition) do parts
   writevtk(Ω,"results",cellfields=["uh"=>uh,"grad_uh"=>∇(uh)])
 end
 ```
-which solves, in parallel, a 2D Poisson problem defined on the unit square. 
-(In order to fully understand the code snippet, familiarity with the high level API of Gridap is assumed.)
[…]
