+++ README.md
@@ -15,7 +15,7 @@ At present, `GridapDistributed.jl` provides scalable parallel data structures fo
 
 ## Documentation
 
-Given that GridapDistributed.jl and Gridap.jl share almost the same high-level API. We refer to the documentation of `Gridap.jl` for more details about the API. In the example below, we show the minor differences among the APIs of `Gridap.jl` and `GridapDistributed.jl`. We refer to the `test/` folder for additional `GridapDistributed.jl` examples.
+Given that `GridapDistributed.jl` and `Gridap.jl` share almost the same high-level API. We refer to the documentation of `Gridap.jl` for more details about the API. In the example below, we show the minor differences among the APIs of `Gridap.jl` and `GridapDistributed.jl`. We refer to the `test/` folder for additional `GridapDistributed.jl` examples.
 
 `GridapDistributed.jl` driver programs can be either run in sequential execution mode--very useful for developing/debugging parallel programs--, see `test/sequential/` folder for examples, or in message-passing (MPI) execution mode--when you want to deploy the code in the actual parallel computer and perform a fast simulation., see `test/mpi/` folder for examples.
 
