+++ docs/make.jl
@@ -8,7 +8,7 @@ makedocs(;
     ],
     repo="https://github.com/gridap/GridapDistributed.jl/blob/{commit}{path}#L{line}",
     sitename="GridapDistributed.jl",
-    authors="Francesc Verdugo <fverdugo@cimne.upc.edu>",
+    authors="S. Badia <santiago.badia@monash.edu>", "A. F. Martin <alberto.martin@monash.edu>", "F. Verdugo <fverdugo@cimne.upc.edu>",
     assets=String[],
 )
 
