### 2022-02-11 cb033f4
Adding skeleton documentation and initial pages
