+++ src/Visualization.jl
@@ -1,3 +1,5 @@
+"""
+"""
 struct DistributedVisualizationData{A<:AbstractPData}
   visdata::A
 end
