+++ docs/src/Visualization.md
@@ -1,0 +1,6 @@
+# Visualization
+
+```@autodocs
+Modules = [GridapDistributed]
+Pages   = ["Visualization.jl"]
+```
