+++ docs/src/FESpaces.md
@@ -1,0 +1,6 @@
+# FESpaces
+
+```@autodocs
+Modules = [GridapDistributed]
+Pages   = ["FESpaces.jl"]
+```
