+++ docs/src/DivConformingFESpaces.md
@@ -1,0 +1,6 @@
+# DivConformingFESpaces
+
+```@autodocs
+Modules = [GridapDistributed]
+Pages   = ["DivConformingFESpaces.jl"]
+```
