+++ docs/src/GridapDistributed.md
@@ -1,0 +1,6 @@
+# GridapDistributed
+
+```@autodocs
+Modules = [GridapDistributed]
+Pages   = ["GridapDistributed.jl"]
+```
