+++ dev/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"getting-started/#Getting-Started-1","page":"Getting Started","title":"Getting Started","text":"","category":"section"},{"location":"getting-started/#Installation-requirements-1","page":"Getting Started","title":"Installation requirements","text":"","category":"section"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"GridapDistributed is tested on Linux, but it should be also possible to use it on Mac OS and Windows since it is written exclusively in Julia and it only depends on registered Julia packages.","category":"page"},{"location":"getting-started/#Installation-1","page":"Getting Started","title":"Installation","text":"","category":"section"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"GridapDistributed is a registered package. Thus, the installation should be straight forward using the Julia's package manager Pkg. To this end, open the Julia REPL (i.e., execute the julia binary), type ] to enter package mode, and install GridapDistributed as follows","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"pkg> add GridapDistributed","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"You will also need the PartitionedArrays package. ","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"pkg> add PartitionedArrays","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"If you want to leverage the satellite packages of GridapDistributed.jl, i.e., GridapPETSc.jl, GridapP4est.jl, and/or GridapGmsh.jl, you have to install them separately as ","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"pkg> add GridapPETSc","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"pkg> add GridapGmsh","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"pkg> add GridapP4est","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"Please note that these three packages depend on binary builds of the corresponding libraries they wrap (i.e., PETSc, Gmsh, and P4est). By default, they will leverage binary builds available at the Julia package registry. However, one may also use custom installations of these libraries. See the documentation of the corresponding package for more details","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"For further information about how to install and manage Julia packages, see the Pkg documentation.","category":"page"},{"location":"getting-started/#Further-steps-1","page":"Getting Started","title":"Further steps","text":"","category":"section"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"If you are new to the Gridap ecosystem of packages, we recommend that you first follow the Gridap Tutorials step by step in order to get familiar with the Gridap.jl library. GridapDistributed.jl and Gridap.jl share almost the same high-level API. Therefore, some familiarity with Gridap.jl is highly recommended (if not essential) before starting with GridapDistributed.jl. ","category":"page"},{"location":"getting-started/#","page":"Getting Started","title":"Getting Started","text":"If you are already familiarized with Gridap.jl, we recommend you to start straight away with the following tutorial.","category":"page"},{"location":"MultiField/#MultiField-1","page":"MultiField","title":"MultiField
[…]
