### 2022-02-11 89ab200
Some missing methods in DistributedDiscreteModels
