### 2022-02-11 ad1121a
Allow visualization of functions and numbers
