### 2022-02-12 7030e01
Fixes for PR #74
