### 2022-02-13 #4157 @whedon
[REVIEW]: GridapDistributed: a massively parallel finite element toolbox in Julia
**Submitting author:** @amartinhuertas (<a href="http://orcid.org/0000-0001-5751-4561">Dr. Alberto F. Martin</a>)
**Repository:** <a href="https://github.com/gridap/GridapDistributed.jl/" target ="_blank">https://github.com/gridap/GridapDistributed.jl/</a>
**Version:** v0.2.4
