### 2022-02-13 4737441
update NEWS.md and .gitignore
