+++ test/sequential/runtests.jl
@@ -16,4 +16,6 @@ using Test
 
 @time @testset "DivConformingTests" begin include("DivConformingTests.jl") end
 
+@time @testset "TransientDistributedCellFieldTests" begin include("TransientDistributedCellFieldTests.jl") end
+
 end # module
