+++ test/sequential/TransientDistributedCellFieldTests.jl
@@ -1,5 +1,5 @@
 module TransientDistributedCellFieldTestsSeq
 using PartitionedArrays
 include("../TransientDistributedCellFieldTests.jl")
-prun(TransientDistributedCellFieldTests.main,sequential,(2,))
+prun(TransientDistributedCellFieldTests.main,sequential,(2,2))
 end # module
