### 2022-02-14 fa9873e
Removed zero_dirichlet_values and associated functions
