### 2022-02-15 6bb5531
Heat equation in mpi tests
