+++ test/GridapODEs/TransientDistributedCellFieldTests.jl
@@ -26,16 +26,16 @@ function main(parts)
   @test isa(dda(0),GridapDistributed.DistributedCellField)
 
   b(t) = TransientCellField(a(t),(da(t),dda(t)))
-  @test isa(b(0),GridapDistributed.TransientDistributedCellField)
-  @test isa(b(0),GridapDistributed.TransientSingleFieldDistributedCellField)
+  @test isa(b(0),GridapDistributed.GridapODEs.TransientDistributedCellField)
+  @test isa(b(0),GridapDistributed.GridapODEs.TransientSingleFieldDistributedCellField)
 
   db(t) = ∂t(b(t))
-  @test isa(db(0),GridapDistributed.TransientDistributedCellField)
-  @test isa(db(0),GridapDistributed.TransientSingleFieldDistributedCellField)
+  @test isa(db(0),GridapDistributed.GridapODEs.TransientDistributedCellField)
+  @test isa(db(0),GridapDistributed.GridapODEs.TransientSingleFieldDistributedCellField)
 
   ddb(t) = ∂t(db(t))
-  @test isa(ddb(0),GridapDistributed.TransientDistributedCellField)
-  @test isa(ddb(0),GridapDistributed.TransientSingleFieldDistributedCellField)
+  @test isa(ddb(0),GridapDistributed.GridapODEs.TransientDistributedCellField)
+  @test isa(ddb(0),GridapDistributed.GridapODEs.TransientSingleFieldDistributedCellField)
 
   @test (∑(∫(a(0.5))dΩ)) ≈ 0.25
   @test (∑(∫(da(0.5))dΩ)) ≈ 1.0
