+++ test/sequential/runtests.jl
@@ -18,16 +18,6 @@ using Test
 
 @time @testset "DivConformingTests" begin include("DivConformingTests.jl") end
 
-@time @testset "TransientDistributedCellFieldTests" begin
-  include("TransientDistributedCellFieldTests.jl")
-end
-
-@time @testset "TransientMultiFieldDistributedCellFieldTests" begin
-  include("TransientMultiFieldDistributedCellFieldTests.jl")
-end
-
-@time @testset "HeatEquation" begin include("HeatEquationTests.jl") end
-
-@time @testset "StokesOpenBoundary" begin include("StokesOpenBoundaryTests.jl") end
+@time @testset "GridapODEs" begin include("GridapODEs/runtests.jl") end
 
 end # module
