+++ src/GridapDistributed.jl
@@ -11,8 +11,6 @@ using Gridap.CellData
 using Gridap.Visualization
 using Gridap.FESpaces
 using Gridap.MultiField
-using GridapODEs.TransientFETools
-using GridapODEs.ODETools
 
 using PartitionedArrays
 const PArrays = PartitionedArrays
@@ -25,7 +23,6 @@ import Gridap.TensorValues: inner, outer, double_contraction, symmetric_part
 import LinearAlgebra: det, tr, cross, dot, ⋅
 import Base: inv, abs, abs2, *, +, -, /, adjoint, transpose, real, imag, conj
 import Gridap.Fields: grad2curl
-import GridapODEs.ODETools: ∂t, ∂tt
 
 export FullyAssembledRows
 export SubAssembledRows
@@ -44,10 +41,6 @@ include("DivConformingFESpaces.jl")
 
 include("MultiField.jl")
 
-include("TransientDistributedCellField.jl")
-
-include("TransientMultiFieldDistributedCellField.jl")
-
-include("TransientFESpaces.jl")
+include("GridapODEs/GridapODEs.jl")
 
 end # module
