+++ test/sequential/GridapODEs/HeatEquationTests.jl
@@ -1,5 +1,5 @@
 module HeatEquationTestsSeq
 using PartitionedArrays
-include("../HeatEquationTests.jl")
+include("../../GridapODEs/HeatEquationTests.jl")
 prun(HeatEquationTests.main,sequential,(2,2))
 end # module
