+++ test/sequential/GridapODEs/TransientMultiFieldDistributedCellFieldTests.jl
@@ -1,5 +1,5 @@
 module TransientMultiFieldDistributedCellFieldTestsSeq
 using PartitionedArrays
-include("../TransientMultiFieldDistributedCellFieldTests.jl")
+include("../../GridapODEs/TransientMultiFieldDistributedCellFieldTests.jl")
 prun(TransientMultiFieldDistributedCellFieldTests.main,sequential,(2,2))
 end # module
