+++ test/sequential/GridapODEs/TransientDistributedCellFieldTests.jl
@@ -1,5 +1,5 @@
 module TransientDistributedCellFieldTestsSeq
 using PartitionedArrays
-include("../TransientDistributedCellFieldTests.jl")
+include("../../GridapODEs/TransientDistributedCellFieldTests.jl")
 prun(TransientDistributedCellFieldTests.main,sequential,(2,2))
 end # module

[…]
