+++ Project.toml
@@ -6,7 +6,6 @@ version = "0.2.5"
 [deps]
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 Gridap = "56d4f2e9-7ea1-5844-9cf6-b9c51ca7ce8e"
-GridapODEs = "55e38337-5b6e-4c7c-9cfc-e00dd49102e6"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
 PartitionedArrays = "5a9dfac6-5c52-46f7-8278-5e2210713be9"
@@ -15,10 +14,10 @@ SparseMatricesCSR = "a0a7dd2c-ebf4-11e9-1f05-cf50bc540ca1"
 WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 
 [compat]
-Gridap = "0.17.8"
-PartitionedArrays = "0.2.10"
 FillArrays = "0.8.4, 0.9, 0.10, 0.11, 0.12"
+Gridap = "0.17.9"
 MPI = "0.16, 0.17, 0.18, 0.19"
+PartitionedArrays = "0.2.10"
 SparseMatricesCSR = "0.6.6"
 WriteVTK = "1.12.0"
 julia = "1.3"
