### 2022-02-28 #4157 @arfon
> @arfon it looks like whedon still does the reminders. ☝️

Yeah, that will stop being the case over the next ~2 weeks. Apologies for any confusion caused.
