### 2022-03-08 #4157 @Kevin-Mattheus-Moerman
@PetrKryslUCSD thanks for your help here! Are you able to tick that last box? The license box is un-ticked. Note the MIT license used is OSI approved and accepted for JOSS.
