+++ test/sequential/ODEs/runtests.jl
@@ -1,4 +1,4 @@
-module GridapODEsTests
+module ODEsTests
 
 using Test
 
