+++ test/sequential/ODEs/TransientDistributedCellFieldTests.jl
@@ -1,5 +1,5 @@
 module TransientDistributedCellFieldTestsSeq
 using PartitionedArrays
-include("../../GridapODEs/TransientDistributedCellFieldTests.jl")
+include("../../ODEs/TransientDistributedCellFieldTests.jl")
 prun(TransientDistributedCellFieldTests.main,sequential,(2,2))
 end # module
