+++ test/ODEs/HeatEquationTests.jl
@@ -1,7 +1,7 @@
 module HeatEquationTests
 
 using Gridap
-using Gridap.GridapODEs
+using Gridap.ODEs
 using GridapDistributed
 using PartitionedArrays
 using Test
