+++ test/sequential/ODEs/HeatEquationTests.jl
@@ -1,5 +1,5 @@
 module HeatEquationTestsSeq
 using PartitionedArrays
-include("../../GridapODEs/HeatEquationTests.jl")
+include("../../ODEs/HeatEquationTests.jl")
 prun(HeatEquationTests.main,sequential,(2,2))
 end # module
