+++ test/sequential/ODEs/TransientMultiFieldDistributedCellFieldTests.jl
@@ -1,5 +1,5 @@
 module TransientMultiFieldDistributedCellFieldTestsSeq
 using PartitionedArrays
-include("../../GridapODEs/TransientMultiFieldDistributedCellFieldTests.jl")
+include("../../ODEs/TransientMultiFieldDistributedCellFieldTests.jl")
 prun(TransientMultiFieldDistributedCellFieldTests.main,sequential,(2,2))
 end # module
