+++ test/sequential/ODEs/StokesOpenBoundaryTests.jl
@@ -1,5 +1,5 @@
 module StokesOpenBoundaryTestsSeq
 using PartitionedArrays
-include("../../GridapODEs/StokesOpenBoundaryTests.jl")
+include("../../ODEs/StokesOpenBoundaryTests.jl")
 prun(StokesOpenBoundaryTests.main,sequential,(2,2))
 end # module
