### 2022-03-11 4b8b08c
Adding surface coupling test (sequential mode)
