+++ test/sequential/runtests.jl
@@ -18,4 +18,6 @@ using Test
 
 @time @testset "DivConformingTests" begin include("DivConformingTests.jl") end
 
+@time @testset "SurfaceCouplingTests" begin include("SurfaceCouplingTests.jl") end
+
 end # module
