### 2022-03-11 d670825
Testing surface coupling also in MPI
