+++ test/sequential/SurfaceCouplingTests.jl
@@ -1,104 +1,6 @@
-module SurfaceCouplingTests
-
-using Gridap
-import Gridap: ∇
-using GridapDistributed
+module SurfaceCouplingTestsSeq
 using PartitionedArrays
-using Test
-
-# Analytical functions
-u(x) = VectorValue( x[1]^2 + 2*x[2]^2, -x[1]^2 )
-∇u(x) = TensorValue( 2*x[1], 4*x[2], -2*x[1], zero(x[1]) )
-Δu(x) = VectorValue( 6, -2 )
-p(x) = x[1] + 3*x[2]
-∇p(x) = VectorValue(1,3)
-s(x) = -Δu(x)
-f(x) = -Δu(x) + ∇p(x)
-g(x) = tr(∇u(x))
-∇(::typeof(u)) = ∇u
-∇(::typeof(p)) = ∇p
-
-# Mesh
-parts = get_part_ids(sequential,(2,2))
-cells = (10,10)
-domain = (0,1,0,1)
-model = CartesianDiscreteModel(parts,domain,cells)
-function is_in(coords)
-  R = 0.4
-  n = length(coords)
-  x = (1/n)*sum(coords) - Point(0.5,0.5)
-  d = x[1]^2 + x[2]^2 - R^2
-  d < 0
-end
-cell_to_entity = map_parts(model.models) do model
-  grid = get_grid(model)
-  cell_to_coords = get_cell_coordinates(grid)
-  cell_to_is_solid = lazy_map(is_in,cell_to_coords)
-  cell_to_is_fluid = lazy_map(!,cell_to_is_solid)
-  labels = get_face_labeling(model)
-  cell_to_entity = labels.d_to_dface_to_entity[end]
-  solid = maximum(cell_to_entity) + 1
-  fluid = solid + 1
-  cell_to_entity[Gridap.Arrays.collect1d(cell_to_is_solid)] .= solid
-  cell_to_entity[Gridap.Arrays.collect1d(cell_to_is_fluid)] .= fluid
-  add_tag!(labels,"solid",[solid])
-  add_tag!(labels,"fluid",[fluid])
-  add_tag_from_tags!(labels,"dirichlet",[1,2,5])
-  add_tag_from_tags!(labels,"neumann",[6,7,8])
-  cell_to_entity
-end
-exchange!(cell_to_entity,model.gids.exchanger) # Make tags consistent
-
-# Domains and measures
-Ω = Interior(model)
-Ωs = Interior(model,tags="solid")
-Ωf = Interior(model,tags="fluid")
-Λ = Boundary(model,tags="neumann")
-Γ = Interface(Ωf,Ωs)
-n_Λ = get_normal_vector(Λ)
-n_Γ = get_normal_vector(Γ)
-k = 2
-degree = 2*k
-dΩ = Measure(Ω,degree)
-dΩs = Measure(Ωs,degree)
-dΩf = Measure(Ωf,degree)
-dΛ = Measure(Λ,degree)
-dΓ = Measure(Γ,degree)
-
-# FE Spaces
-reffe_u = ReferenceFE(lagrangian,VectorValue{2,Float64},k)
-reffe_p = ReferenceFE(lagrangian,Float64,k-1,space=:P)
-V = TestFESpace(Ω,reffe_u,dirichlet_tags="dirichlet")
-Q = TestFESpace(Ωf,reffe_p)
[…]
