+++ test/sequential/runtests.jl
@@ -18,6 +18,16 @@ using Test
 
 @time @testset "DivConformingTests" begin include("DivConformingTests.jl") end
 
-@time @testset "ODEs" begin include("ODEs/runtests.jl") end
+@time @testset "TransientDistributedCellFieldTests" begin
+  include("TransientDistributedCellFieldTests.jl")
+end
+
+@time @testset "TransientMultiFieldDistributedCellFieldTests" begin
+  include("TransientMultiFieldDistributedCellFieldTests.jl")
+end
+
+@time @testset "HeatEquation" begin include("HeatEquationTests.jl") end
+
+@time @testset "StokesOpenBoundary" begin include("StokesOpenBoundaryTests.jl") end
 
 end # module
