+++ test/sequential/TransientMultiFieldDistributedCellFieldTests.jl
@@ -1,5 +1,5 @@
 module TransientMultiFieldDistributedCellFieldTestsSeq
 using PartitionedArrays
-include("../../ODEs/TransientMultiFieldDistributedCellFieldTests.jl")
+include("../TransientMultiFieldDistributedCellFieldTests.jl")
 prun(TransientMultiFieldDistributedCellFieldTests.main,sequential,(2,2))
 end # module
