+++ test/sequential/HeatEquationTests.jl
@@ -1,5 +1,5 @@
 module HeatEquationTestsSeq
 using PartitionedArrays
-include("../../ODEs/HeatEquationTests.jl")
+include("../HeatEquationTests.jl")
 prun(HeatEquationTests.main,sequential,(2,2))
 end # module
