+++ test/TransientMultiFieldDistributedCellFieldTests.jl
@@ -35,28 +35,28 @@ function main(parts)
   @test isa(dda(0),GridapDistributed.DistributedMultiFieldFEFunction)
 
   b(t) = TransientCellField(a(t),(da(t),dda(t)))
-  @test isa(b(0),GridapDistributed.ODEs.TransientDistributedCellField)
-  @test isa(b(0),GridapDistributed.ODEs.TransientMultiFieldDistributedCellField)
+  @test isa(b(0),GridapDistributed.TransientDistributedCellField)
+  @test isa(b(0),GridapDistributed.TransientMultiFieldDistributedCellField)
 
   db(t) = ∂t(b(t))
-  @test isa(db(0),GridapDistributed.ODEs.TransientDistributedCellField)
-  @test isa(db(0),GridapDistributed.ODEs.TransientMultiFieldDistributedCellField)
+  @test isa(db(0),GridapDistributed.TransientDistributedCellField)
+  @test isa(db(0),GridapDistributed.TransientMultiFieldDistributedCellField)
 
   ddb(t) = ∂t(db(t))
-  @test isa(ddb(0),GridapDistributed.ODEs.TransientDistributedCellField)
-  @test isa(ddb(0),GridapDistributed.ODEs.TransientMultiFieldDistributedCellField)
+  @test isa(ddb(0),GridapDistributed.TransientDistributedCellField)
+  @test isa(ddb(0),GridapDistributed.TransientMultiFieldDistributedCellField)
 
   b1(t) = b(t)[1]
-  @test isa(b1(0),GridapDistributed.ODEs.TransientDistributedCellField)
-  @test isa(b1(0),GridapDistributed.ODEs.TransientSingleFieldDistributedCellField)
+  @test isa(b1(0),GridapDistributed.TransientDistributedCellField)
+  @test isa(b1(0),GridapDistributed.TransientSingleFieldDistributedCellField)
 
   db1(t) = ∂t(b1(t))
-  @test isa(db1(0),GridapDistributed.ODEs.TransientDistributedCellField)
-  @test isa(db1(0),GridapDistributed.ODEs.TransientSingleFieldDistributedCellField)
+  @test isa(db1(0),GridapDistributed.TransientDistributedCellField)
+  @test isa(db1(0),GridapDistributed.TransientSingleFieldDistributedCellField)
 
   ddb1(t) = ∂t(db1(t))
-  @test isa(ddb1(0),GridapDistributed.ODEs.TransientDistributedCellField)
-  @test isa(ddb1(0),GridapDistributed.ODEs.TransientSingleFieldDistributedCellField)
+  @test isa(ddb1(0),GridapDistributed.TransientDistributedCellField)
+  @test isa(ddb1(0),GridapDistributed.TransientSingleFieldDistributedCellField)
 
   @test (∑(∫(b(0.5)[1])dΩ)) == (∑(∫(b1(0.5))dΩ))
   @test (∑(∫(db(0.5)[1])dΩ)) == (∑(∫(db1(0.5))dΩ))
