+++ Project.toml
@@ -14,7 +14,7 @@ SparseMatricesCSR = "a0a7dd2c-ebf4-11e9-1f05-cf50bc540ca1"
 WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
 
 [compat]
-Gridap = "0.17.8"
+Gridap = "0.17.11"
 PartitionedArrays = "0.2.10"
 FillArrays = "0.8.4, 0.9, 0.10, 0.11, 0.12"
 MPI = "0.16, 0.17, 0.18, 0.19"
