### 2022-03-25 #4157 @Leticia-maria
I would like to be a permanent reviewer of JOSS, is it possible?
