### 2022-03-26 e5b846a
First approach to support FE spaces defined on lower dimensional triangulations
