+++ src/Geometry.jl
@@ -302,7 +302,8 @@ end
 
 function Geometry.Triangulation(
   model::DistributedDiscreteModel;kwargs...)
-  Triangulation(no_ghost,model;kwargs...)
+  D=num_cell_dims(model)
+  Triangulation(no_ghost,ReferenceFE{D},model;kwargs...)
 end
 
 function Geometry.BoundaryTriangulation(
@@ -316,11 +317,26 @@ function Geometry.SkeletonTriangulation(
 end
 
 function Geometry.Triangulation(
-  portion,model::DistributedDiscreteModel;kwargs...)
+  portion,::Type{ReferenceFE{D}},model::DistributedDiscreteModel{D};kwargs...) where {D}
   trians = map_parts(model.models,model.gids.partition) do model,gids
-    Triangulation(portion,gids,model;kwargs...)
+     Triangulation(portion,gids,ReferenceFE{D},model;kwargs...)
   end
   DistributedTriangulation(trians,model)
+end
+
+function Geometry.Triangulation(
+  portion,::Type{ReferenceFE{Dt}},model::DistributedDiscreteModel{Dm};kwargs...) where {Dt,Dm}
+  trians = map_parts(model.models) do model
+     Triangulation(ReferenceFE{Dt},model;kwargs...)
+  end
+  dtrian=DistributedTriangulation(trians,model)
+
+  # Generate global ordering for the objects of dimension Dt
+  tgids=generate_cell_gids(dtrian)
+  trians = map_parts(model.models,tgids.partition) do model, gids
+    Triangulation(portion,gids,ReferenceFE{Dt},model;kwargs...)
+  end
+  dtrian=DistributedTriangulation(trians,model)
 end
 
 function Geometry.BoundaryTriangulation(
@@ -340,7 +356,7 @@ function Geometry.SkeletonTriangulation(
 end
 
 function Geometry.Triangulation(
-  portion,gids::AbstractIndexSet,args...;kwargs...)
+  portion,gids::AbstractIndexSet, args...;kwargs...) where D
   trian = Triangulation(args...;kwargs...)
   filter_cells_when_needed(portion,gids,trian)
 end
@@ -367,6 +383,17 @@ function Geometry.InterfaceTriangulation(a::DistributedTriangulation,b::Distribu
   trians = map_parts(InterfaceTriangulation,a.trians,b.trians)
   @assert a.model === b.model
   DistributedTriangulation(trians,a.model)
+end
+
+function Geometry.Triangulation(
+  portion, model::DistributedDiscreteModel;kwargs...)
+  D=num_cell_dims(model)
+  Triangulation(portion,ReferenceFE{D},model;kwargs...)
+end
+
+function Geometry.Triangulation(
+  ::Type{ReferenceFE{D}}, model::DistributedDiscreteModel;kwargs...) where D
+  Triangulation(no_ghost, ReferenceFE{D}, model; kwargs...)
 end
 
 function filter_cells_when_needed(
@@ -410,9 +437,17 @@ end
 # defined on the boundary (e.g. a Lagrange multiplier)
 function remove_ghost_cells(trian,gids)
   model = get_background_model(trian)
-  D = num_cell_dims(model)
-  glue = get_glue(trian,Val(D))
+  Dm    = num_cell_dims(model)
+  Dt    = num_cell_dims(trian)
+  glue = get_glue(trian,Val(Dm))
   remove_ghost_cells(glue,trian,gids)
+end
+
+function remove_ghost_cells(glue::Nothing,trian,gids)
+  @notimplementedif num_cells(trian)!=length(gids.lid_to_part)
+  mcell_to_part = gids.lid_to_part
+  tcell_to_mask = mcell_to_part .== gids.part
[…]
