### 2022-04-29 #4157 @Kevin-Mattheus-Moerman
@amartinhuertas thanks for your comment. Our apologies for the delays here. It has been very difficult across the board (since the COVID pandemic), for all JOSS submissions, to find reviewers and also to get them to commit to a timely review. We've been forced to be more flexible regarding time as a consequence. Having said that I think here it has been taking a bit long for them to respond so I have just contacted the reviewers separately just now again, and hopefully we'll hear from them shortly.

In the unlikely case the reviewers become unresponsive then yes we can find different reviewers, however lets wait to give them a change to get back to us. Thanks for your patience.
