### 2022-05-16 4cbb828
Added zero_dirichlet_values
