### 2022-05-24 #4157 @jedbrown
Sorry about my delays (a medical emergency last week laid waste to all my plans). Thanks for this cool package.

## Project docs
1. I know it's hard to address disparate audiences, but I think it's pretty unclear to someone used to invoking mpiexec themselves what the workflow looks like. Could you address such users somewhere discoverable, including instructions for running a *.jl demo (in parallel) rather than mainly pointing to notebook workflows?
1. I'm honestly surprised the shapes broadcast here. In any case, this isn't the Jacobian of the previous line, which represents a regularized p=4 Laplacian. https://github.com/gridap/GridapDistributed.jl/blob/master/test/PLaplacianTests.jl#L27

## Paper
1. "compact syntax, that": remove comma
1. use @key instead of [@key] for [in-text citation](https://pandoc.org/MANUAL.html#extension-citations), as with "see (Verdugo & Badia, 2021) for a detailed overview" (and subsequent occurrences)
1. the code sample will render correctly if you put it in a fenced Julia block, and that way will use a font that matches the other literals in the text. I know JOSS failed to handle unicode fonts properly in older iterations, but this was fixed a few months ago.
1. Not a problem per se, but I'm surprised to see the same time step used for the 256 (approx 40 km) and 512 (approx 20 km) elements. I'd expect to step at a constant CFL (otherwise you won't resolve phase). With a 480s time step, this would resolve waves on the order of 80 m/s versus 40 m/s on the two meshes, respectively. The latter is slower than the jet stream. Also, I'm not sure what to think of the ringing artifacts in the (low resolution) figure 3a.
1. Many references have have improper capitalization (APA says to use sentence case)
1. Gibson et al has missing journal info
1. Discussion of related packages is slim; I'd suggest at least discussing relative scope of FEniCS and perhaps something like FreeFEM++ and MFEM/deal.II/libMesh.
