+++ joss_paper/paper.bib
@@ -22,16 +22,19 @@
     month  = jun
 }
 
+
+
 @article{Verdugo:2021,
-  	url = {https://arxiv.org/abs/2109.12818},
-  	Archiveprefix = {arXiv},
-  	Author = {{Verdugo}, F. and {Badia}, S.},
-  	Eprint = {2109.12818},
-  	Journal = {ArXiv e-prints},
-  	Keywords = {Mathematical software},
-  	Month = sep,
-  	Title = {{The software design of Gridap: a Finite Element package based on the Julia JIT compiler}},
-  	Year = 2021
+  doi = {10.1016/j.cpc.2022.108341},
+  url = {https://doi.org/10.1016/j.cpc.2022.108341},
+  year = {2022},
+  month = jul,
+  publisher = {Elsevier {BV}},
+  volume = {276},
+  pages = {108341},
+  author = {Francesc Verdugo and Santiago Badia},
+  title = {The software design of {G}ridap: a finite element package based on the {J}ulia {JIT} compiler},
+  journal = {Computer Physics Communications}
 }
 
 @misc{gridapetsc,
@@ -75,7 +78,7 @@ month = {feb},
 number = {1},
 pages = {65--98},
 publisher = {Society for Industrial and Applied Mathematics},
-title = {{Julia: A Fresh Approach to Numerical Computing}},
+title = {{Julia: a fresh approach to numerical computing}},
 volume = {59},
 year = {2017}
 }
@@ -90,7 +93,7 @@ month = {aug},
 number = {52},
 pages = {2520},
 publisher = {The Open Journal},
-title = {{Gridap: An extensible Finite Element toolbox in Julia}},
+title = {{Gridap: an extensible finite element toolbox in Julia}},
 url = {https://joss.theoj.org/papers/10.21105/joss.02520},
 volume = {5},
 year = {2020}
@@ -109,18 +112,18 @@ month = {dec},
 number = {6},
 pages = {C436--C468},
 publisher = {Society for Industrial and Applied Mathematics},
-title = {{A Generic Finite Element Framework on Parallel Tree-Based Adaptive Meshes}},
+title = {{A generic finite element framework on parallel tree-based adaptive meshes}},
 volume = {42},
 year = {2020}
 }
-@article{Gibson2019,
+@book{Gibson2019,
 address = {Cham},
 author = {Gibson, Thomas H. and McRae, Andrew T.T. and Cotter, Colin J. and Mitchell, Lawrence and Ham, David A.},
 doi = {10.1007/978-3-030-23957-2},
 isbn = {978-3-030-23956-5},
 publisher = {Springer International Publishing},
 series = {Mathematics of Planet Earth},
-title = {{Compatible Finite Element Methods for Geophysical Flows}},
+title = {{Compatible finite element methods for geophysical flows}},
 url = {http://link.springer.com/10.1007/978-3-030-23957-2},
 year = {2019}
 }
@@ -149,7 +152,7 @@ month = {may},
 number = {3},
 pages = {1103--1133},
 publisher = {Society for Industrial and Applied Mathematics},
-title = {{p4est: Scalable Algorithms for Parallel Adaptive Mesh Refinement on Forests of Octrees}},
+title = {{p4est: scalable algorithms for parallel adaptive mesh refinement on forests of octrees}},
 volume = {33},
 year = {2011}
 }
@@ -163,7 +166,7 @@ month = {mar},
 number = {1},
 pages = {93--114},
 publisher = {Academic Press},
-title = {{The “Cubed Sphere”: A New Method for the Solution of Partial Differential Equations in Spherical Geometry}},
+title = {{The “Cubed Sphere”: a new method for the solution of partial differential equations in spherical geometry}},
 volume = {124},
[…]
