### 2022-05-26 #4157 @amartinhuertas
Hi @jedbrown,

thanks for your review.

I have pushed a new version of the article which I believe addresses all your concerns. Let me know if this is not the case.

Let me answer below inline to each comment.

> I know it's hard to address disparate audiences, but I think it's pretty unclear to someone used to invoking mpiexec themselves what the workflow looks like. Could you address such users somewhere discoverable, including instructions for running a *.jl demo (in parallel) rather than mainly pointing to notebook workflows?

Yes. See https://github.com/gridap/GridapDistributed.jl#execution-modes-and-how-to-execute-the-program-in-each-mode

> I'm honestly surprised the shapes broadcast here. In any case, this isn't the Jacobian of the previous line, which represents a regularized p=4 Laplacian. https://github.com/gridap/GridapDistributed.jl/blob/master/test/PLaplacianTests.jl#L27

I am not sure which is the source of the confusion (if there is any), nor why this is not the Jacobian of the previous line. Can you please elaborate a little bit more? Please note that `∇du` represents the gradient applied to the trial basis, while  `∇u` the gradient applied to a single FE function which respresents the current nonlinear solver iterate.

1. Addressed in the new version of the article.
2. Addressed in the new version of the article.
3. I tried the Julia fenced block but still Unicode characterers are not rendered appropiately. I am used the Docker image supplied with joss to render the paper in my local machine.
4. Addressed in the new version of the article.
5. Addressed in the new version of the article.
6. It is a book. I have fixed the bib entry.
7. Addressed in the new version of the article.

@amartinhuertas
