+++ joss_paper/paper.md
@@ -41,8 +41,7 @@ This equips end-users with a tool for the development of simulation codes able t
 In order to scale FE simulations to large core counts, the mesh used to discretize the computational domain on which the PDE is posed must be partitioned (distributed) among the parallel tasks such that each of these only holds a local portion of the global mesh. The same requirement applies to the rest of data structures in the FE simulation pipeline, i.e., FE space, linear system, solvers, data output, etc. The local portion of each task is composed by a set of cells that it owns, i.e., the *local cells* of the task, and a set of off-processor cells (owned by remote processors) which are in touch with its local cells, i.e., the *ghost cells* of the task [@Badia2020a].
 This overlapped mesh partition is used by GridapDistributed, among others, to exchange data among nearest neighbors, and to glue together global Degrees of Freedom (DoFs) which are sitting on the interface among subdomains. Following this design principle, GridapDistributed provides scalable parallel data structures and associated methods for simple grid handling (in particular, Cartesian-like meshes of arbitrary-dimensional, topologically n-cube domains), FE spaces setup, and distributed linear system assembly. It is in our future plans to provide highly scalable linear and nonlinear solvers tailored for the FE discretization of PDEs (e.g., linear and nonlinear matrix-free geometric multigrid and domain decomposition preconditioners). In the meantime, however, GridapDistributed can be combined with other Julia packages in order to realize the full potential required in real-world applications. These packages and their relation with GridapDistributed are overviewed in the next section. 
 
-There are a number of highly quality open source parallel finite element packages available in the literature. Some examples are deal.II [@dealII93], libMesh [@libMeshPaper],
-, MFEM [@mfem], FEMPAR [@Badia2017], FEniCS [@Kirby2006], or FreeFEM++ [@freefem], to name a few. All these  packages have their own set of features, potentials, and limitations. Among these, FEniCS and FreeFEM++ are perhaps the closest ones in scope and spirit to the packages in the Gridap ecosystem. A hallmark of the latter ones is that they are able to achieve a very expressive and compact (yet efficient) syntax without relying on a sophisticated compiler of variational forms
+There are a number of high quality open source parallel finite element packages available in the literature. Some examples are deal.II [@dealII93], libMesh [@libMeshPaper], MFEM [@mfem], FEMPAR [@Badia2017], FEniCS [@Kirby2006], or FreeFEM++ [@freefem], to name a few. All these  packages have their own set of features, potentials, and limitations. Among these, FEniCS and FreeFEM++ are perhaps the closest ones in scope and spirit to the packages in the Gridap ecosystem. A hallmark of the latter ones is that they are able to achieve a very expressive and compact (yet efficient) syntax without relying on a sophisticated compiler of variational forms
 (see, e.g., @Kirby2006). To this end, this high level syntax is transformed into low-level code using the Julia JIT compiler. Apart from a more intrincate workflow (e.g., a Python front-end and a C/C++ back-end), one of the limitations of form compilers is that they are typically quite rigid systems not designed to be extended by average users.
 
 
