### 2022-05-26 #4157 @jedbrown
Regarding the 4-Laplacian, the current code is
```julia
  σ(∇u) =(1.0+∇u⋅∇u)*∇u
  dσ(∇du,∇u) = 2*∇u⋅∇du + (1.0+∇u⋅∇u)*∇du
```
The first term of the second line is a scalar while the second is a vector. In Julia, we normally can't add these without `.+` broadcast; e.g., `1 + [2,3]` is an error and you need `1 .+ [2,3]`.

I usually write the derivative like this
```julia
  dσ(∇du,∇u) = (2*∇u⋅∇du)*∇u + (1.0+∇u⋅∇u)*∇du
```
or something like this to really expose the geometric intuition that Newton linearization yields a symmetric rank-1 modification to the isotropic effective diffusivity.
```julia
  μ(∇u) = 1 + ∇u⋅∇u
  σ(∇u) = μ(∇u) * ∇u
  dσ(∇du,∇u) = (μ(∇u)*I + 2 * ∇u * ∇u') * ∇du
```
(In 3D, $p>2$ elongates it like a pencil while $p<2$ squashes it like a pancake because the second term has a negative multiplier. The problem is singular in the $p=1$ limit and degenerate in the $p=∞$ limit.)

> I tried the Julia fenced block but still Unicode characterers are not rendered appropiately. I am using the Docker image supplied with joss to render the paper in my local machine.

Hmm, I ran like this and it seems okay?
```julia
$ docker run -it --rm -v $(pwd):/data -u $(id -u):$(id -g) openjournals/inara -o pdf,crossref joss_paper/paper.md
```

![image](https://user-images.githubusercontent.com/3303/170489803-e65fe302-0f93-47d9-9400-570409638d66.png)
```diff
diff --git i/joss_paper/paper.md w/joss_paper/paper.md
index 8f8a49d..ff3998a 100644
--- i/joss_paper/paper.md
+++ w/joss_paper/paper.md
@@ -55,7 +55,31 @@ In order to confirm our previous claims on expressiveness, conciseness and produ
 (In order to fully understand the code snippet, familiarity with the high level API of Gridap is assumed.)
 The domain is discretized using the parallel Cartesian-like mesh generator built-in in GridapDistributed. The only minimal burden posed on the programmer versus Gridap is a call to the `prun` function of PartitionedArrays right at the beginning of the program. With this function, the programmer sets up the PartitionedArrays communication backend (i.e., MPI communication backend in the example), specifies the number of parts and their layout (i.e., `(2,2)` 2D layout in the example), and provides a function (using Julia do-block syntax for function arguments in the example) to be run on each part. This function is equivalent to a sequential Gridap script, except for the `CartesianDiscreteModel` call, which, in GridapDistributed, also requires the `parts` argument passed back by the `prun` function. In a typical cluster environment, this example would be executed on 4 MPI tasks from a terminal as `mpirun -n 4 julia --project=. example.jl`.

-![](code.pdf)
+```julia
+using Gridap
+using GridapDistributed
+using PartitionedArrays
+partition = (2,2)
+prun(mpi,partition) do parts
+  domain = (0,1,0,1)
+  mesh_partition = (4,4)
+  model = CartesianDiscreteModel(parts,domain,mesh_partition)
+  order = 2
+  u((x,y)) = (x+y)^order
+  f(x) = -Δ(u,x)
+  reffe = ReferenceFE(lagrangian,Float64,order)
+  V = TestFESpace(model,reffe,dirichlet_tags="boundary")
+  U = TrialFESpace(u,V)
+  Ω = Triangulation(model)
+  dΩ = Measure(Ω,2*order)
+  a(u,v) = ∫( ∇(v)·∇(u) )dΩ
+  l(v) = ∫( v*f )dΩ
+  op = AffineFEOperator(a,l,U,V)
+  uh = solve(op)
+  writevtk(Ω,"results",cellfields=["uh"=>uh,"grad_uh"=>∇(uh)])
+end
+```
+

 # Parallel scaling benchmark
```
