### 2022-05-26 #4157 @amartinhuertas
> I usually write the derivative like this

Ok. I know see what you mean. You are absolutely correct. I wrote the derivative a thousand times on paper and I was sure that it matched the code, but in the code there was a missing multiplication by `∇u` in the first term! Pair programming is definitely efficient, I think I would have not been able to see this error in 1000 years.

The code snippet below reproduces the situation. The last assert fails meaning that, as one would expect, `2*∇u⋅∇du` is NOT equivalent to `(2*∇u⋅∇du)*∇u`. In regards to why it broadcasts (i.e., why no error is produced), the cause is that operations among Gridap `CellField`s (e.g., `∇u` and `∇du` are `CellField`s) have broadcasting semantics by construction. Once an operation tree among `CellField`s is built,  this tree is then evaluated at a set of integration points during numerical integration. The operations among tensors of floating point numbers which are triggered are broadcast-like operations.

```julia
using Gridap

domain = (0,1,0,1)
cells = (4,4)
model = CartesianDiscreteModel(domain,cells)
Ω  = Interior(model)
dΩ = Measure(Ω,2)

reffe = ReferenceFE(lagrangian,Float64,1)
V = TestFESpace(model,reffe,dirichlet_tags="boundary")
U = TrialFESpace(V)

du  = get_trial_fe_basis(U)
∇du = gradient(du)

u = FEFunction(U,rand(num_free_dofs(U)))
∇u = gradient(u)

op1=2*∇u⋅∇du
op2=(1.0+∇u⋅∇u)*∇du
op3=(2*∇u⋅∇du)*∇u
x=get_cell_points(dΩ.quad)

op1(x)[1]
op2(x)[1]
op3(x)[1]
@assert (op1+op2)(x)[1] == op1(x)[1] .+ op2(x)[1]
@assert (op2+op3)(x)[1] == (op1+op2)(x)[1]
```
