+++ joss_paper/code.tex
@@ -1,11 +1,0 @@
-%\pdfminorversion=4
-\documentclass[11pt]{minimal}
-% Packages
-\usepackage[utf8x]{inputenc}
-\usepackage[usenames,dvipsnames,svgnames]{xcolor}
-\usepackage{textgreek}
-\usepackage{minted}
-\definecolor{bg}{rgb}{0.93,0.93,0.93}
-\begin{document}
-\inputminted{julia1}{Example.jl}
-\end{document}
