### 2022-05-26 #4157 @amartinhuertas
> Hmm, I ran like this and it seems okay?

Ok. I tried your patch + suggested command and now it renders correctly. Thanks! I would swear that the command I was using was taken from the JOSS instructions per authors. But I now see that https://joss.readthedocs.io/en/latest/submitting.html?highlight=docker#docker does not match the command I was using. Nevermind, it works :-)
