### 2022-05-26 #4157 @jedbrown
Yeah, JOSS switched to the new software stack (which contained https://github.com/openjournals/inara/pull/4 to handle unicode fonts, plus PDF/A support and other benefits) a couple days after you first submitted. So no mistake on your end, but we can use the new system now (and it'll have better accessibility).
