### 2022-06-07 #4157 @jedbrown
@amartinhuertas Are you going to update the p-Laplacian test?

I think the FEniCS book (https://fenicsproject.org/book/) is a better reference. I think this needs some rewording because it seems the point is that FEniCS *does* rely on a (reasonably sophisticated) form compiler that is problematic to apply to various challenging problems. For example, return mapping algorithms in plasticity need to be written in C++ with bits specific to dolfin/firedrake, thus losing many of the benefits of working with UFL in the first place. The pragmatic need for such escape hatches is why the likes of [PyOP2](https://github.com/OP2/PyOP2) were created.

> A hallmark of the latter ones is that they are able to achieve a very expressive and compact (yet efficient) syntax without relying on a sophisticated compiler of variational forms (see, e.g., @Kirby2006). To this end, this high level syntax is transformed into low-level code using the Julia JIT compiler. Apart from a more intrincate workflow (e.g., a Python front-end and a C/C++ back-end), one of the limitations of form compilers is that they are typically quite rigid systems not designed to be extended by average users.
