### 2022-06-07 30c2920
DistributedDiscreteModel is now able to handle on demand global IDs of lower dimensional faces
