+++ src/Geometry.jl
@@ -77,12 +77,14 @@ end
 """
 struct DistributedDiscreteModel{Dc,Dp,A,B} <: GridapType
   models::A
-  gids::B
+  face_gids::B
   function DistributedDiscreteModel(
     models::AbstractPData{<:DiscreteModel{Dc,Dp}}, gids::PRange) where {Dc,Dp}
     A = typeof(models)
-    B = typeof(gids)
-    new{Dc,Dp,A,B}(models,gids)
+    face_gids=Vector{PRange}(undef,Dc+1)
+    face_gids[Dc+1]=gids
+    B = typeof(face_gids)
+    new{Dc,Dp,A,B}(models,face_gids)
   end
 end
 
@@ -93,8 +95,37 @@ Geometry.num_cell_dims(::Type{<:DistributedDiscreteModel{Dc,Dp}}) where {Dc,Dp} 
 Geometry.num_point_dims(::DistributedDiscreteModel{Dc,Dp}) where {Dc,Dp} = Dp
 Geometry.num_point_dims(::Type{<:DistributedDiscreteModel{Dc,Dp}}) where {Dc,Dp} = Dp
 
-function Geometry.num_cells(model::DistributedDiscreteModel)
- num_gids(model.gids)
+function Geometry.num_cells(model::DistributedDiscreteModel{Dc}) where Dc
+ num_gids(model.face_gids[Dc+1])
+end
+
+function Geometry.num_facets(model::DistributedDiscreteModel{Dc}) where Dc
+  _setup_face_gids!(model,Dc-1)
+  num_gids(model.face_gids[Dc])
+end
+
+function Geometry.num_edges(model::DistributedDiscreteModel{Dc}) where Dc
+  _setup_face_gids!(model,1)
+  num_gids(model.face_gids[2])
+end
+
+function Geometry.num_vertices(model::DistributedDiscreteModel{Dc}) where Dc
+  _setup_face_gids!(model,0)
+  num_gids(model.face_gids[1])
+end
+
+function Geometry.num_faces(model::DistributedDiscreteModel{Dc},dim::Integer) where Dc
+  _setup_face_gids!(model,dim)
+  num_gids(model.face_gids[dim+1])
+end
+
+function Geometry.num_faces(model::DistributedDiscreteModel{Dc}) where Dc
+  num_faces=0
+  for dim=0:Dc
+    _setup_face_gids!(model,dim)
+    num_faces+=num_gids(model.face_gids[dim+1])
+  end
+  num_faces
 end
 
 function Geometry.get_grid(model::DistributedDiscreteModel)
@@ -107,6 +138,15 @@ end
 
 function Geometry.get_face_labeling(model::DistributedDiscreteModel)
   DistributedFaceLabeling(map_parts(get_face_labeling,model.models))
+end
+
+function get_cell_gids(model::DistributedDiscreteModel{Dc}) where Dc
+  model.face_gids[Dc+1]
+end
+
+function get_face_gids(model::DistributedDiscreteModel,dim::Integer)
+  _setup_face_gids!(model,dim)
+  num_gids(model.face_gids[dim+1])
 end
 
 # CartesianDiscreteModel
@@ -318,7 +358,7 @@ end
 
 function Geometry.Triangulation(
   portion,::Type{ReferenceFE{D}},model::DistributedDiscreteModel{D};kwargs...) where {D}
-  trians = map_parts(model.models,model.gids.partition) do model,gids
+  trians = map_parts(model.models,model.face_gids[D+1].partition) do model,gids
      Triangulation(portion,gids,ReferenceFE{D},model;kwargs...)
   end
   DistributedTriangulation(trians,model)
@@ -340,16 +380,16 @@ function Geometry.Triangulation(
 end
 
 function Geometry.BoundaryTriangulation(
[…]
