### 2022-06-07 #4157 @amartinhuertas
> Am I missing something else here? (i.e., another source file with the same error?)

Ok, I think I know now what happened. I updated the master branch, but you were looking at the code in the joss_paper branch. I have now pull master into joss_paper branch so that both branches are synced.
