+++ joss_paper/paper.bib
@@ -210,6 +210,7 @@ year = {2016}
 }
 
 @article {freefem,
+    doi    = {10.1515/jnum-2012-0013},
     AUTHOR = {Hecht, F.}, TITLE = {New development in {FreeFem++}},
    JOURNAL = {J. Numer. Math.},  FJOURNAL = {Journal of Numerical Mathematics},
     VOLUME = {20}, YEAR = {2012},
@@ -232,6 +233,7 @@ year = {2016}
 }
 
 @Article{libMeshPaper,
+  doi     = {10.1007/s00366-006-0049-3},
   author  = {B.~S.~Kirk and J.~W.~Peterson and R.~H.~Stogner and G.~F.~Carey},
   title   = {{\texttt{libMesh}: A C++ library for parallel adaptive mesh refinement/coarsening simulations}},
   journal = {Engineering with Computers},
@@ -276,4 +278,13 @@ pages = {417--444},
 title = {{A compiler for variational forms}},
 volume = {32},
 year = {2006}
+}
+
+@book{fenics-book,
+  doi = {10.1007/978-3-642-23099-8},
+  url = {https://doi.org/10.1007/978-3-642-23099-8},
+  year = {2012},
+  publisher = {Springer Berlin Heidelberg},
+  editor = {Anders Logg and Kent-Andre Mardal and Garth Wells},
+  title = {Automated Solution of Differential Equations by the Finite Element Method}
 }
