### 2022-06-07 fc60672
Rephrasing sentence
