### 2022-06-07 #4157 @amartinhuertas
> Also, the libMesh paper is https://doi.org/10.1007/s00366-006-0049-3

Done!
