### 2022-06-07 defc252
Pending changes in Geometry.jl to support spaces on lower dim trians
