+++ src/Geometry.jl
@@ -146,7 +146,7 @@ end
 
 function get_face_gids(model::DistributedDiscreteModel,dim::Integer)
   _setup_face_gids!(model,dim)
-  num_gids(model.face_gids[dim+1])
+  model.face_gids[dim+1]
 end
 
 # CartesianDiscreteModel
@@ -357,23 +357,10 @@ function Geometry.SkeletonTriangulation(
 end
 
 function Geometry.Triangulation(
-  portion,::Type{ReferenceFE{D}},model::DistributedDiscreteModel{D};kwargs...) where {D}
-  trians = map_parts(model.models,model.face_gids[D+1].partition) do model,gids
-     Triangulation(portion,gids,ReferenceFE{D},model;kwargs...)
-  end
-  DistributedTriangulation(trians,model)
-end
-
-function Geometry.Triangulation(
   portion,::Type{ReferenceFE{Dt}},model::DistributedDiscreteModel{Dm};kwargs...) where {Dt,Dm}
-  trians = map_parts(model.models) do model
-     Triangulation(ReferenceFE{Dt},model;kwargs...)
-  end
-  dtrian=DistributedTriangulation(trians,model)
-
-  # Generate global ordering for the objects of dimension Dt
-  tgids=generate_cell_gids(dtrian)
-  trians = map_parts(model.models,tgids.partition) do model, gids
+  # Generate global ordering for the faces of dimension Dt (if needed)
+  gids=get_face_gids(model,Dt)
+  trians = map_parts(model.models,gids.partition) do model, gids
     Triangulation(portion,gids,ReferenceFE{Dt},model;kwargs...)
   end
   dtrian=DistributedTriangulation(trians,model)
@@ -468,26 +455,18 @@ function filter_cells_when_needed(
   remove_ghost_cells(trian,cell_gids)
 end
 
-# For the moment remove_ghost_cells
-# refers to the triangulation faces
-# pointing into the ghost cells
-# in the underlying background discrete
-# model. This might change when solving
-# multi-field PDEs with one of the fields
-# defined on the boundary (e.g. a Lagrange multiplier)
-function remove_ghost_cells(trian,gids)
+function remove_ghost_cells(trian::Triangulation,gids)
   model = get_background_model(trian)
-  Dm    = num_cell_dims(model)
   Dt    = num_cell_dims(trian)
-  glue = get_glue(trian,Val(Dm))
+  glue = get_glue(trian,Val(Dt))
   remove_ghost_cells(glue,trian,gids)
 end
 
-function remove_ghost_cells(glue::Nothing,trian,gids)
-  @notimplementedif num_cells(trian)!=length(gids.lid_to_part)
-  mcell_to_part = gids.lid_to_part
-  tcell_to_mask = mcell_to_part .== gids.part
-  view(trian, findall(tcell_to_mask))
+function remove_ghost_cells(trian::Union{SkeletonTriangulation,BoundaryTriangulation},gids)
+  model = get_background_model(trian)
+  Dm    = num_cell_dims(model)
+  glue = get_glue(trian,Val(Dm))
+  remove_ghost_cells(glue,trian,gids)
 end
 
 function remove_ghost_cells(glue::FaceToFaceGlue,trian,gids)
@@ -536,30 +515,42 @@ function add_ghost_cells(dtrian::DistributedTriangulation)
   add_ghost_cells(dmodel,dtrian)
 end
 
-function add_ghost_cells(dmodel::DistributedDiscreteModel{Dm},
-                         dtrian::DistributedTriangulation{Dm}) where Dm
[…]
