### 2022-06-07 #4157 @jedbrown
> There are a number of high quality open source parallel finite element packages available in the literature. Some examples are deal.II [@dealII93], libMesh [@libMeshPaper], MFEM [@mfem], FEMPAR [@Badia2017], FEniCS [@fenics-book], or FreeFEM++ [@freefem], to name a few. All these  packages have their own set of features, potentials, and limitations. Among these, FEniCS and FreeFEM++ are perhaps the closest ones in scope and spirit to the packages in the Gridap ecosystem. A hallmark of the latter ones compared to the former ones is that their very expressive and compact (yet efficient) syntax is transformed into low-level code using the Julia JIT compiler and thus they do not need a sophisticated compiler of variational forms nor a more intrincate workflow (e.g., a Python front-end and a C/C++ back-end).

1. This seems to be saying that a hallmark of FEniCS/FreeFEM++ relative to deal.II et al is that it uses Julia JIT. I think your intent is perhaps that a hallmark of Gridap relative to these is that you implement weak forms in a full-featured language (Julia) and thus benefit from Julia's JIT rather than a custom one.
2. "intrincate" -> "intricate"
