### 2022-06-07 0e71a72
Update NEWS.md
