+++ NEWS.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased 
+## [0.2.6] - 2022-06-07
 
 ### Added
 - Support for parallel ODE solvers (GridapDistributed+GridapODEs). Since PR [#81](https://github.com/gridap/GridapDistributed.jl/pull/81)
