+++ Project.toml
@@ -1,7 +1,7 @@
 name = "GridapDistributed"
 uuid = "f9701e48-63b3-45aa-9a63-9bc6c271f355"
 authors = ["S. Badia <santiago.badia@monash.edu>", "A. F. Martin <alberto.martin@monash.edu>", "F. Verdugo <fverdugo@cimne.upc.edu>"]
-version = "0.2.5"
+version = "0.2.6"
 
 [deps]
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
