### 2022-06-09 #4157 @amartinhuertas
> The review started off for version v0.2.4, can you confirm if this version is still accurate or if we need to update the version tag for this submission?

The latest public release of GridapDistributed.jl is v0.2.6. We need to update the version tag for this submission to v0.2.6.

> Can you please archive a copy of the reviewed work on ZENODO? You can do this manually if you like. Some users find [these automated steps](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) useful.

Done.

> Please ensure that the ZENODO title and authors match those of the paper.

Done.

> Can you please list the ZENODO archive DOI here?

10.5281/zenodo.6622081
