### 2022-06-09 cba5b33
Use single image per figure with markdown syntax
