### 2022-06-09 #4157 @xuanxu
The [code for a couple of figures](https://github.com/gridap/GridapDistributed.jl/blob/joss_paper/joss_paper/paper.md?plain=1#L91-L95) with multiple images with the same caption was invalid. [This PR](https://github.com/gridap/GridapDistributed.jl/pull/90) changes it to use a single image with multiple panels and using [valid syntax](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).
