### 2022-06-09 #4157 @Kevin-Mattheus-Moerman
@xuanxu this does not explain the reference checking to behave differently though right? :point_up:
